Created a new script to make the products reorder outside the content manager
