+++ Assets/Scripts/OthersControllers/ReOrderProductsController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 84a92d87676295741b43d553c6b4fba7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
